Addition of raycasting script
Additions:
Raycasting script added to player (main camera). This script will be used to mainly interact with objects such as buttons. Currently bugged, pushing to Github to have a backup of the original script.
Invisible object where the towers are, to make the raycast have a larger object to hit.
Bug fixes/Updates:
None.